# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/feed/comment": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "评论动态",
                "parameters": [
                    {
                        "description": "动态ID、评论内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CommentPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"评论详情",
                        "schema": {
                            "$ref": "#/definitions/entity.CommentResp"
                        }
                    }
                }
            }
        },
        "/feed/comment_delete": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "删除评论",
                "parameters": [
                    {
                        "description": "动态ID、主评论ID、子评论ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.DeleteCommentPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
        "/feed/comment_list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取评论列表",
                "parameters": [
                    {
                        "type": "string",
                        "name": "feed_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"评论列表",
                        "schema": {
                            "$ref": "#/definitions/entity.GetCommentListResp"
                        }
                    }
                }
            }
        },
        "/feed/comment_update": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "更新评论",
                "parameters": [
                    {
                        "description": "动态ID、主评论ID、子评论ID、评论内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UpdateCommentPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"更新后的评论详情",
                        "schema": {
                            "$ref": "#/definitions/entity.CommentInfo"
                        }
                    }
                }
            }
        },
        "/feed/info": {
            "get": {
                "description": "使用但不校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "获取动态详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "动态ID",
                        "name": "feed_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"动态详情",
                        "schema": {
                            "$ref": "#/definitions/entity.GetFeedInfoResp"
                        }
                    }
                }
            }
        },
        "/feed/like": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "点赞|取消点赞 动态",
                "parameters": [
                    {
                        "description": "动态ID、点赞或取消点赞",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.LikeFeedPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
        "/feed/list": {
            "get": {
                "description": "使用但不校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "搜索动态",
                "parameters": [
                    {
                        "type": "string",
                        "name": "author_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "keywords",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"动态列表",
                        "schema": {
                            "$ref": "#/definitions/entity.GetFeedListResp"
                        }
                    }
                }
            }
        },
        "/feed/related_feeds": {
            "get": {
                "description": "使用但不校验登录态。",
                "produces": [
                    "application/json"
                ],
                "summary": "获取同一空间下相关动态列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "space_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"动态列表",
                        "schema": {
                            "$ref": "#/definitions/entity.GetRelatedFeedsResp"
                        }
                    }
                }
            }
        },
        "/feed/reply": {
            "post": {
                "description": "校验登录态",
                "produces": [
                    "application/json"
                ],
                "summary": "回复评论",
                "parameters": [
                    {
                        "description": "动态ID、评论ID、目标用户ID、回复内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ReplyPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" \"回复评论详情",
                        "schema": {
                            "$ref": "#/definitions/entity.ReplyInfo"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Author": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
        "entity.CommentInfo": {
            "type": "object",
            "properties": {
                "commentator": {
                    "$ref": "#/definitions/entity.Commentator"
                },
                "content": {
                    "type": "string"
                },
                "creation_time": {
                    "type": "string"
                },
                "feed_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "sub_comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.ReplyInfo"
                    }
                },
                "update_time": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "entity.CommentPayload": {
            "type": "object",
            "required": [
                "content",
                "feed_id"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "feed_id": {
                    "type": "string"
                }
            }
        },
        "entity.CommentResp": {
            "type": "object",
            "properties": {
                "commentator": {
                    "$ref": "#/definitions/entity.Commentator"
                },
                "content": {
                    "type": "string"
                },
                "creation_time": {
                    "type": "string"
                },
                "feed_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "sub_comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.ReplyInfo"
                    }
                },
                "update_time": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "entity.Commentator": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
...
        "entity.Creator": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
        "entity.DeleteCommentPayload": {
            "type": "object",
            "required": [
                "feed_id"
            ],
            "properties": {
                "comment_id": {
                    "type": "string"
                },
                "feed_id": {
                    "type": "string"
                },
                "sub_comment_id": {
                    "type": "string"
                }
            }
        },
...
        "entity.DocInfo": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/entity.Author"
                },
                "author_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "cover": {
                    "type": "string"
                },
                "creation_time": {
                    "type": "string"
                },
                "doc_id": {
                    "type": "string"
                },
                "public": {
                    "type": "boolean"
                },
                "space_id": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                }
            }
        },
        "entity.FeedInfo": {
            "type": "object",
            "properties": {
                "collected": {
                    "type": "boolean"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Comment"
                    }
                },
                "comments_count": {
                    "type": "integer"
                },
                "creation_time": {
                    "type": "string"
                },
                "creator": {
                    "$ref": "#/definitions/entity.Creator"
                },
                "creator_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "likes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.LikeInfo"
                    }
                },
                "likes_count": {
                    "type": "integer"
                },
                "subject": {
                    "$ref": "#/definitions/entity.DocInfo"
                },
                "subject_id": {
                    "type": "string"
                },
                "subject_type": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                }
            }
        },
...
        "entity.GetCommentListResp": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.CommentInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "entity.GetFeedInfoResp": {
            "type": "object",
            "properties": {
                "collected": {
                    "type": "boolean"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Comment"
                    }
                },
                "comments_count": {
                    "type": "integer"
                },
                "creation_time": {
                    "type": "string"
                },
                "creator": {
                    "$ref": "#/definitions/entity.Creator"
                },
                "creator_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "likes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.LikeInfo"
                    }
                },
                "likes_count": {
                    "type": "integer"
                },
                "subject": {
                    "$ref": "#/definitions/entity.DocInfo"
                },
                "subject_id": {
                    "type": "string"
                },
                "subject_type": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                }
            }
        },
        "entity.GetFeedListResp": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.FeedInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "entity.GetRelatedFeedsResp": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.FeedInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "entity.LikeFeedPayload": {
            "type": "object",
            "required": [
                "event",
                "feed_id"
            ],
            "properties": {
                "event": {
                    "type": "string"
                },
                "feed_id": {
                    "type": "string"
                }
            }
        },
        "entity.LikeInfo": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "creation_time": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "entity.ReplyInfo": {
            "type": "object",
            "properties": {
                "commentator": {
                    "$ref": "#/definitions/entity.Commentator"
                },
                "content": {
                    "type": "string"
                },
                "creation_time": {
                    "type": "string"
                },
                "feed_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "reply_commentator": {
                    "$ref": "#/definitions/entity.Commentator"
                },
                "reply_user_id": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "entity.ReplyPayload": {
            "type": "object",
            "required": [
                "content",
                "feed_id"
            ],
            "properties": {
                "comment_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "feed_id": {
                    "type": "string"
                },
                "reply_user_id": {
                    "type": "string"
                }
            }
        },
...
        "entity.UpdateCommentPayload": {
            "type": "object",
            "required": [
                "feed_id"
            ],
            "properties": {
                "comment_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "feed_id": {
                    "type": "string"
                },
                "sub_comment_id": {
                    "type": "string"
                }
            }
        },
...
        "model.Comment": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "creation_time": {
                    "type": "string"
                },
                "feed_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "sub_comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SubComment"
                    }
                },
                "update_time": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "model.SubComment": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "creation_time": {
                    "type": "string"
                },
                "feed_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "reply_user_id": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    }
}